fix quantity - set scale 8

+++ src/main/java/com/app/financial/service/asset/IAssetCalculator.java
@@ -8,5 +8,5 @@ public interface IAssetCalculator {
 
     BigDecimal indexCalculate(BigDecimal quotation, BigDecimal averagePrice);
 
-    BigDecimal valueCalculate(Integer quantity, BigDecimal quotation);
+    BigDecimal valueCalculate(BigDecimal quantity, BigDecimal quotation);
 }
